+++ library/table.go
@@ -22,11 +22,7 @@ func openTable(s rbxmk.State) *lua.LTable {
 
 func tableClear(s rbxmk.State) int {
 	t := s.L.CheckTable(1)
-	// TODO: Implement native clear function.
-	t.ForEach(func(k, v lua.LValue) error {
-		t.RawSet(k, lua.LNil)
-		return nil
-	})
+	t.Clear()
 	return 0
 }